Add home component with exchange rate display
This commit introduces a new 'home' component displaying the latest exchange rates. The data is supplied by a newly implemented ExchangeRateService, which fetches currency information from an API. The component and its associated html template provide a table view of the exchange rates, allowing for an overview of currency values.