Add error handling for fetching exchange rates
The changes in the code add error handling for the process of fetching historical exchange rates. When an error occurs in the HTTP request for the rates, an error message is now displayed on the page. Along with this, the historical rates list will be emptied, thus, avoiding any display of stale or erroneous data.